% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{batch_create_partition}
\alias{batch_create_partition}
\title{Creates one or more partitions in a batch operation}
\usage{
batch_create_partition(DatabaseName, TableName, PartitionInputList,
  CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the metadata database in which the partition is to be created.}

\item{TableName}{The name of the metadata table in which the partition is to be created.}

\item{PartitionInputList}{A list of \code{PartitionInput} structures that define the partitions to be created.}

\item{CatalogId}{The ID of the catalog in which the partion is to be created. Currently, this should be the AWS account ID.}
}
\description{
Creates one or more partitions in a batch operation.
}
\examples{

}
