% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{batch_delete_partition}
\alias{batch_delete_partition}
\title{Deletes one or more partitions in a batch operation}
\usage{
batch_delete_partition(DatabaseName, TableName, PartitionsToDelete,
  CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database in which the table in question resides.}

\item{TableName}{The name of the table where the partitions to be deleted is located.}

\item{PartitionsToDelete}{A list of \code{PartitionInput} structures that define the partitions to be deleted.}

\item{CatalogId}{The ID of the Data Catalog where the partition to be deleted resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Deletes one or more partitions in a batch operation.
}
\examples{

}
