% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{batch_delete_table_version}
\alias{batch_delete_table_version}
\title{Deletes a specified batch of versions of a table}
\usage{
batch_delete_table_version(DatabaseName, TableName, VersionIds,
  CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The database in the catalog in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{TableName}{The name of the table. For Hive compatibility, this name is entirely lowercase.}

\item{VersionIds}{A list of the IDs of versions to be deleted.}

\item{CatalogId}{The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.}
}
\description{
Deletes a specified batch of versions of a table.
}
\examples{

}
