% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_classifier}
\alias{create_classifier}
\title{Creates a classifier in the user's account}
\usage{
create_classifier(GrokClassifier = NULL, XMLClassifier = NULL,
  JsonClassifier = NULL)
}
\arguments{
\item{GrokClassifier}{A \code{GrokClassifier} object specifying the classifier to create.}

\item{XMLClassifier}{An \code{XMLClassifier} object specifying the classifier to create.}

\item{JsonClassifier}{A \code{JsonClassifier} object specifying the classifier to create.}
}
\description{
Creates a classifier in the user's account. This may be a \code{GrokClassifier}, an \code{XMLClassifier}, or abbrev \code{JsonClassifier}, depending on which field of the request is present.
}
\section{Accepted Parameters}{
\preformatted{create_classifier(
  GrokClassifier = list(
    Classification = "string",
    Name = "string",
    GrokPattern = "string",
    CustomPatterns = "string"
  ),
  XMLClassifier = list(
    Classification = "string",
    Name = "string",
    RowTag = "string"
  ),
  JsonClassifier = list(
    Name = "string",
    JsonPath = "string"
  )
)
}
}

