% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_connection}
\alias{create_connection}
\title{Creates a connection definition in the Data Catalog}
\usage{
create_connection(CatalogId = NULL, ConnectionInput)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which to create the connection. If none is supplied, the AWS account ID is used by default.}

\item{ConnectionInput}{[required] A \code{ConnectionInput} object defining the connection to create.}
}
\description{
Creates a connection definition in the Data Catalog.
}
\section{Accepted Parameters}{
\preformatted{create_connection(
  CatalogId = "string",
  ConnectionInput = list(
    Name = "string",
    Description = "string",
    ConnectionType = "JDBC"|"SFTP",
    MatchCriteria = list(
      "string"
    ),
    ConnectionProperties = list(
      "string"
    ),
    PhysicalConnectionRequirements = list(
      SubnetId = "string",
      SecurityGroupIdList = list(
        "string"
      ),
      AvailabilityZone = "string"
    )
  )
)
}
}

