% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_crawler}
\alias{create_crawler}
\title{Creates a new crawler with specified targets, role, configuration, and optional schedule}
\usage{
create_crawler(Name, Role, DatabaseName, Targets, Description = NULL,
  Schedule = NULL, Classifiers = NULL, TablePrefix = NULL,
  SchemaChangePolicy = NULL, Configuration = NULL,
  CrawlerSecurityConfiguration = NULL)
}
\arguments{
\item{Name}{Name of the new crawler.}

\item{Role}{The IAM role (or ARN of an IAM role) used by the new crawler to access customer resources.}

\item{DatabaseName}{The AWS Glue database where results are written, such as: \code{arn:aws:daylight:us-east-1::database/sometable/*}.}

\item{Targets}{A list of collection of targets to crawl.}

\item{Description}{A description of the new crawler.}

\item{Schedule}{A \code{cron} expression used to specify the schedule (see \href{http://docs.aws.amazon.com/glue/latest/dg/monitor-data-warehouse-schedule.html}{Time-Based Schedules for Jobs and Crawlers}. For example, to run something every day at 12:15 UTC, you would specify: \code{cron(15 12 * * ? *)}.}

\item{Classifiers}{A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.}

\item{TablePrefix}{The table prefix used for catalog tables that are created.}

\item{SchemaChangePolicy}{Policy for the crawler's update and deletion behavior.}

\item{Configuration}{Crawler configuration information. This versioned JSON string allows users to specify aspects of a Crawler's behavior.

You can use this field to force partitions to inherit metadata such as classification, input format, output format, serde information, and schema from their parent table, rather than detect this information separately for each partition. Use the following JSON string to specify that behavior:

Example: \code{'{ "Version": 1.0, "CrawlerOutput": { "Partitions": { "AddOrUpdateBehavior": "InheritFromTable" } } }'}}

\item{CrawlerSecurityConfiguration}{The name of the SecurityConfiguration structure to be used by this Crawler.}
}
\description{
Creates a new crawler with specified targets, role, configuration, and optional schedule. At least one crawl target must be specified, in the \emph{s3Targets} field, the \emph{jdbcTargets} field, or the \emph{DynamoDBTargets} field.
}
\examples{

}
