% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_job}
\alias{create_job}
\title{Creates a new job definition}
\usage{
create_job(Name, Role, Command, Description = NULL, LogUri = NULL,
  ExecutionProperty = NULL, DefaultArguments = NULL, Connections = NULL,
  MaxRetries = NULL, AllocatedCapacity = NULL, Timeout = NULL,
  NotificationProperty = NULL, SecurityConfiguration = NULL)
}
\arguments{
\item{Name}{The name you assign to this job definition. It must be unique in your account.}

\item{Role}{The name or ARN of the IAM role associated with this job.}

\item{Command}{The JobCommand that executes this job.}

\item{Description}{Description of the job being defined.}

\item{LogUri}{This field is reserved for future use.}

\item{ExecutionProperty}{An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.}

\item{DefaultArguments}{The default arguments for this job.

You can specify arguments here that your own job-execution script consumes, as well as arguments that AWS Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the \href{http://docs.aws.amazon.com/glue/latest/dg/aws-glue-programming-python-calling.html}{Calling AWS Glue APIs in Python} topic in the developer guide.

For information about the key-value pairs that AWS Glue consumes to set up your job, see the \href{http://docs.aws.amazon.com/glue/latest/dg/aws-glue-programming-etl-glue-arguments.html}{Special Parameters Used by AWS Glue} topic in the developer guide.}

\item{Connections}{The connections used for this job.}

\item{MaxRetries}{The maximum number of times to retry this job if it fails.}

\item{AllocatedCapacity}{The number of AWS Glue data processing units (DPUs) to allocate to this Job. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the \href{https://aws.amazon.com/glue/pricing/}{AWS Glue pricing page}.}

\item{Timeout}{The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters \code{TIMEOUT} status. The default is 2,880 minutes (48 hours).}

\item{NotificationProperty}{Specifies configuration properties of a job notification.}

\item{SecurityConfiguration}{The name of the SecurityConfiguration structure to be used with this job.}
}
\description{
Creates a new job definition.
}
\examples{

}
