% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_partition}
\alias{create_partition}
\title{Creates a new partition}
\usage{
create_partition(DatabaseName, TableName, PartitionInput, CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the metadata database in which the partition is to be created.}

\item{TableName}{The name of the metadata table in which the partition is to be created.}

\item{PartitionInput}{A \code{PartitionInput} structure defining the partition to be created.}

\item{CatalogId}{The ID of the catalog in which the partion is to be created. Currently, this should be the AWS account ID.}
}
\description{
Creates a new partition.
}
\examples{

}
