% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_table}
\alias{create_table}
\title{Creates a new table definition in the Data Catalog}
\usage{
create_table(DatabaseName, TableInput, CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The catalog database in which to create the new table. For Hive compatibility, this name is entirely lowercase.}

\item{TableInput}{The \code{TableInput} object that defines the metadata table to create in the catalog.}

\item{CatalogId}{The ID of the Data Catalog in which to create the \code{Table}. If none is supplied, the AWS account ID is used by default.}
}
\description{
Creates a new table definition in the Data Catalog.
}
\examples{

}
