% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_classifier}
\alias{delete_classifier}
\title{Removes a classifier from the Data Catalog}
\usage{
delete_classifier(Name)
}
\arguments{
\item{Name}{[required] Name of the classifier to remove.}
}
\description{
Removes a classifier from the Data Catalog.
}
\section{Accepted Parameters}{
\preformatted{delete_classifier(
  Name = "string"
)
}
}

