% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_crawler}
\alias{delete_crawler}
\title{Removes a specified crawler from the Data Catalog, unless the crawler state is \code{RUNNING}}
\usage{
delete_crawler(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] Name of the crawler to remove.}
}
\description{
Removes a specified crawler from the Data Catalog, unless the crawler state is \code{RUNNING}.
}
\examples{

}
