% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_database}
\alias{delete_database}
\title{Removes a specified Database from a Data Catalog}
\usage{
delete_database(Name, CatalogId = NULL)
}
\arguments{
\item{Name}{The name of the Database to delete. For Hive compatibility, this must be all lowercase.}

\item{CatalogId}{The ID of the Data Catalog in which the database resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Removes a specified Database from a Data Catalog.
}
\details{
After completing this operation, you will no longer have access to the tables (and all table versions and partitions that might belong to the tables) and the user-defined functions in the deleted database. AWS Glue deletes these "orphaned" resources asynchronously in a timely manner, at the discretion of the service.

To ensure immediate deletion of all related resources, before calling \code{DeleteDatabase}, use \code{DeleteTableVersion} or \code{BatchDeleteTableVersion}, \code{DeletePartition} or \code{BatchDeletePartition}, \code{DeleteUserDefinedFunction}, and \code{DeleteTable} or \code{BatchDeleteTable}, to delete any resources that belong to the database.
}
\examples{

}
