% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_job}
\alias{delete_job}
\title{Deletes a specified job definition}
\usage{
delete_job(
  JobName = "string"
)
}
\arguments{
\item{JobName}{[required] The name of the job definition to delete.}
}
\description{
Deletes a specified job definition. If the job definition is not found, no exception is thrown.
}
\examples{

}
