% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_resource_policy}
\alias{delete_resource_policy}
\title{Deletes a specified policy}
\usage{
delete_resource_policy(
  PolicyHashCondition = "string"
)
}
\arguments{
\item{PolicyHashCondition}{The hash value returned when this policy was set.}
}
\description{
Deletes a specified policy.
}
\examples{

}
