% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_table}
\alias{delete_table}
\title{Removes a table definition from the Data Catalog}
\usage{
delete_table(DatabaseName, Name, CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{Name}{The name of the table to be deleted. For Hive compatibility, this name is entirely lowercase.}

\item{CatalogId}{The ID of the Data Catalog where the table resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Removes a table definition from the Data Catalog.
}
\details{
After completing this operation, you will no longer have access to the table versions and partitions that belong to the deleted table. AWS Glue deletes these "orphaned" resources asynchronously in a timely manner, at the discretion of the service.

To ensure immediate deletion of all related resources, before calling \code{DeleteTable}, use \code{DeleteTableVersion} or \code{BatchDeleteTableVersion}, and \code{DeletePartition} or \code{BatchDeletePartition}, to delete any resources that belong to the table.
}
\examples{

}
