% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_table_version}
\alias{delete_table_version}
\title{Deletes a specified version of a table}
\usage{
delete_table_version(DatabaseName, TableName, VersionId, CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The database in the catalog in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{TableName}{The name of the table. For Hive compatibility, this name is entirely lowercase.}

\item{VersionId}{The ID of the table version to be deleted.}

\item{CatalogId}{The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.}
}
\description{
Deletes a specified version of a table.
}
\examples{

}
