% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_trigger}
\alias{delete_trigger}
\title{Deletes a specified trigger}
\usage{
delete_trigger(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the trigger to delete.}
}
\description{
Deletes a specified trigger. If the trigger is not found, no exception is thrown.
}
\examples{

}
