% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_catalog_import_status}
\alias{get_catalog_import_status}
\title{Retrieves the status of a migration operation}
\usage{
get_catalog_import_status(CatalogId = NULL)
}
\arguments{
\item{CatalogId}{The ID of the catalog to migrate. Currently, this should be the AWS account ID.}
}
\description{
Retrieves the status of a migration operation.
}
\section{Accepted Parameters}{
\preformatted{get_catalog_import_status(
  CatalogId = "string"
)
}
}

