% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_classifier}
\alias{get_classifier}
\title{Retrieve a classifier by name}
\usage{
get_classifier(Name)
}
\arguments{
\item{Name}{[required] Name of the classifier to retrieve.}
}
\description{
Retrieve a classifier by name.
}
\section{Accepted Parameters}{
\preformatted{get_classifier(
  Name = "string"
)
}
}

