% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_classifiers}
\alias{get_classifiers}
\title{Lists all classifier objects in the Data Catalog}
\usage{
get_classifiers(
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{MaxResults}{Size of the list to return (optional).}

\item{NextToken}{An optional continuation token.}
}
\description{
Lists all classifier objects in the Data Catalog.
}
\examples{

}
