% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_connection}
\alias{get_connection}
\title{Retrieves a connection definition from the Data Catalog}
\usage{
get_connection(CatalogId = NULL, Name)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connection resides. If none is supplied, the AWS account ID is used by default.}

\item{Name}{[required] The name of the connection definition to retrieve.}
}
\description{
Retrieves a connection definition from the Data Catalog.
}
\section{Accepted Parameters}{
\preformatted{get_connection(
  CatalogId = "string",
  Name = "string"
)
}
}

