% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_connections}
\alias{get_connections}
\title{Retrieves a list of connection definitions from the Data Catalog}
\usage{
get_connections(CatalogId = NULL, Filter = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connections reside. If none is supplied, the AWS account ID is used by default.}

\item{Filter}{A filter that controls which connections will be returned.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum number of connections to return in one response.}
}
\description{
Retrieves a list of connection definitions from the Data Catalog.
}
\examples{

}
