% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_crawler}
\alias{get_crawler}
\title{Retrieves metadata for a specified crawler}
\usage{
get_crawler(Name)
}
\arguments{
\item{Name}{Name of the crawler to retrieve metadata for.}
}
\description{
Retrieves metadata for a specified crawler.
}
\examples{

}
