% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_crawler_metrics}
\alias{get_crawler_metrics}
\title{Retrieves metrics about specified crawlers}
\usage{
get_crawler_metrics(
  CrawlerNameList = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{CrawlerNameList}{A list of the names of crawlers about which to retrieve metrics.}

\item{MaxResults}{The maximum size of a list to return.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Retrieves metrics about specified crawlers.
}
\examples{

}
