% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_crawlers}
\alias{get_crawlers}
\title{Retrieves metadata for all crawlers defined in the customer account}
\usage{
get_crawlers(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of crawlers to return on each call.}

\item{NextToken}{A continuation token, if this is a continuation request.}
}
\description{
Retrieves metadata for all crawlers defined in the customer account.
}
\examples{

}
