% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_data_catalog_encryption_settings}
\alias{get_data_catalog_encryption_settings}
\title{Retrieves the security configuration for a specified catalog}
\usage{
get_data_catalog_encryption_settings(CatalogId = NULL)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog for which to retrieve the security configuration. If none is supplied, the AWS account ID is used by default.}
}
\description{
Retrieves the security configuration for a specified catalog.
}
\section{Accepted Parameters}{
\preformatted{get_data_catalog_encryption_settings(
  CatalogId = "string"
)
}
}

