% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_dataflow_graph}
\alias{get_dataflow_graph}
\title{Transforms a Python script into a directed acyclic graph (DAG)}
\usage{
get_dataflow_graph(PythonScript = NULL)
}
\arguments{
\item{PythonScript}{The Python script to transform.}
}
\description{
Transforms a Python script into a directed acyclic graph (DAG).
}
\examples{

}
