% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_dev_endpoint}
\alias{get_dev_endpoint}
\title{Retrieves information about a specified DevEndpoint}
\usage{
get_dev_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{Name of the DevEndpoint for which to retrieve information.}
}
\description{
Retrieves information about a specified DevEndpoint.
}
\examples{

}
