% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_job}
\alias{get_job}
\title{Retrieves an existing job definition}
\usage{
get_job(
  JobName = "string"
)
}
\arguments{
\item{JobName}{[required] The name of the job definition to retrieve.}
}
\description{
Retrieves an existing job definition.
}
\examples{

}
