% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_job_runs}
\alias{get_job_runs}
\title{Retrieves metadata for all runs of a given job definition}
\usage{
get_job_runs(JobName, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{JobName}{The name of the job definition for which to retrieve all job runs.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum size of the response.}
}
\description{
Retrieves metadata for all runs of a given job definition.
}
\examples{

}
