% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_jobs}
\alias{get_jobs}
\title{Retrieves all current job definitions}
\usage{
get_jobs(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum size of the response.}
}
\description{
Retrieves all current job definitions.
}
\examples{

}
