% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_mapping}
\alias{get_mapping}
\title{Creates mappings}
\usage{
get_mapping(Source, Sinks = NULL, Location = NULL)
}
\arguments{
\item{Source}{Specifies the source table.}

\item{Sinks}{A list of target tables.}

\item{Location}{Parameters for the mapping.}
}
\description{
Creates mappings.
}
\examples{

}
