% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_partition}
\alias{get_partition}
\title{Retrieves information about a specified partition}
\usage{
get_partition(DatabaseName, TableName, PartitionValues, CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database where the partition resides.}

\item{TableName}{The name of the partition's table.}

\item{PartitionValues}{The values that define the partition.}

\item{CatalogId}{The ID of the Data Catalog where the partition in question resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Retrieves information about a specified partition.
}
\examples{

}
