% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_partitions}
\alias{get_partitions}
\title{Retrieves information about the partitions in a table}
\usage{
get_partitions(DatabaseName, TableName, CatalogId = NULL, Expression = NULL,
  NextToken = NULL, Segment = NULL, MaxResults = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database where the partitions reside.}

\item{TableName}{The name of the partitions' table.}

\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If none is supplied, the AWS account ID is used by default.}

\item{Expression}{An expression filtering the partitions to be returned.}

\item{NextToken}{A continuation token, if this is not the first call to retrieve these partitions.}

\item{Segment}{The segment of the table's partitions to scan in this request.}

\item{MaxResults}{The maximum number of partitions to return in a single response.}
}
\description{
Retrieves information about the partitions in a table.
}
\examples{

}
