% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_plan}
\alias{get_plan}
\title{Gets code to perform a specified mapping}
\usage{
get_plan(Mapping, Source, Sinks = NULL, Location = NULL, Language = NULL)
}
\arguments{
\item{Mapping}{The list of mappings from a source table to target tables.}

\item{Source}{The source table.}

\item{Sinks}{The target tables.}

\item{Location}{Parameters for the mapping.}

\item{Language}{The programming language of the code to perform the mapping.}
}
\description{
Gets code to perform a specified mapping.
}
\examples{

}
