% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_resource_policy}
\alias{get_resource_policy}
\title{Retrieves a specified resource policy}
\usage{
get_resource_policy()
}
\description{
Retrieves a specified resource policy.
}
\examples{

}
