% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_security_configuration}
\alias{get_security_configuration}
\title{Retrieves a specified security configuration}
\usage{
get_security_configuration(Name)
}
\arguments{
\item{Name}{[required] The name of the security configuration to retrieve.}
}
\description{
Retrieves a specified security configuration.
}
\section{Accepted Parameters}{
\preformatted{get_security_configuration(
  Name = "string"
)
}
}

