% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_table_version}
\alias{get_table_version}
\title{Retrieves a specified version of a table}
\usage{
get_table_version(DatabaseName, TableName, CatalogId = NULL,
  VersionId = NULL)
}
\arguments{
\item{DatabaseName}{The database in the catalog in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{TableName}{The name of the table. For Hive compatibility, this name is entirely lowercase.}

\item{CatalogId}{The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.}

\item{VersionId}{The ID value of the table version to be retrieved.}
}
\description{
Retrieves a specified version of a table.
}
\examples{

}
