% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_table_versions}
\alias{get_table_versions}
\title{Retrieves a list of strings that identify available versions of a specified table}
\usage{
get_table_versions(DatabaseName, TableName, CatalogId = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{DatabaseName}{The database in the catalog in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{TableName}{The name of the table. For Hive compatibility, this name is entirely lowercase.}

\item{CatalogId}{The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.}

\item{NextToken}{A continuation token, if this is not the first call.}

\item{MaxResults}{The maximum number of table versions to return in one response.}
}
\description{
Retrieves a list of strings that identify available versions of a specified table.
}
\examples{

}
