% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_trigger}
\alias{get_trigger}
\title{Retrieves the definition of a trigger}
\usage{
get_trigger(Name)
}
\arguments{
\item{Name}{The name of the trigger to retrieve.}
}
\description{
Retrieves the definition of a trigger.
}
\examples{

}
