% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_user_defined_functions}
\alias{get_user_defined_functions}
\title{Retrieves a multiple function definitions from the Data Catalog}
\usage{
get_user_defined_functions(DatabaseName, Pattern, CatalogId = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database where the functions are located.}

\item{Pattern}{An optional function-name pattern string that filters the function definitions returned.}

\item{CatalogId}{The ID of the Data Catalog where the functions to be retrieved are located. If none is supplied, the AWS account ID is used by default.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum number of functions to return in one response.}
}
\description{
Retrieves a multiple function definitions from the Data Catalog.
}
\examples{

}
