% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{import_catalog_to_glue}
\alias{import_catalog_to_glue}
\title{Imports an existing Athena Data Catalog to AWS Glue}
\usage{
import_catalog_to_glue(CatalogId = NULL)
}
\arguments{
\item{CatalogId}{The ID of the catalog to import. Currently, this should be the AWS account ID.}
}
\description{
Imports an existing Athena Data Catalog to AWS Glue
}
\examples{

}
