% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_package.R
\docType{package}
\name{paws.glue-package}
\alias{paws.glue}
\alias{paws.glue-package}
\title{paws.glue: AWS Glue}
\description{
Defines the public endpoint for the AWS Glue
    service.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
