% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{reset_job_bookmark}
\alias{reset_job_bookmark}
\title{Resets a bookmark entry}
\usage{
reset_job_bookmark(JobName)
}
\arguments{
\item{JobName}{The name of the job in question.}
}
\description{
Resets a bookmark entry.
}
\examples{

}
