% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{start_crawler}
\alias{start_crawler}
\title{Starts a crawl using the specified crawler, regardless of what is scheduled}
\usage{
start_crawler(Name)
}
\arguments{
\item{Name}{Name of the crawler to start.}
}
\description{
Starts a crawl using the specified crawler, regardless of what is scheduled. If the crawler is already running, returns a \href{https://docs.aws.amazon.com/glue/latest/dg/aws-glue-api-exceptions.html#aws-glue-api-exceptions-CrawlerRunningException}{CrawlerRunningException}.
}
\examples{

}
