% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{start_crawler_schedule}
\alias{start_crawler_schedule}
\title{Changes the schedule state of the specified crawler to \code{SCHEDULED}, unless the crawler is already running or the schedule state is already \code{SCHEDULED}}
\usage{
start_crawler_schedule(CrawlerName)
}
\arguments{
\item{CrawlerName}{Name of the crawler to schedule.}
}
\description{
Changes the schedule state of the specified crawler to \code{SCHEDULED}, unless the crawler is already running or the schedule state is already \code{SCHEDULED}.
}
\examples{

}
