% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{stop_crawler}
\alias{stop_crawler}
\title{If the specified crawler is running, stops the crawl}
\usage{
stop_crawler(Name)
}
\arguments{
\item{Name}{[required] Name of the crawler to stop.}
}
\description{
If the specified crawler is running, stops the crawl.
}
\section{Accepted Parameters}{
\preformatted{stop_crawler(
  Name = "string"
)
}
}

