% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{stop_crawler_schedule}
\alias{stop_crawler_schedule}
\title{Sets the schedule state of the specified crawler to NOT_SCHEDULED, but does not stop the crawler if it is already running}
\usage{
stop_crawler_schedule(CrawlerName)
}
\arguments{
\item{CrawlerName}{[required] Name of the crawler whose schedule state to set.}
}
\description{
Sets the schedule state of the specified crawler to \code{NOT_SCHEDULED}, but does not stop the crawler if it is already running.
}
\section{Accepted Parameters}{
\preformatted{stop_crawler_schedule(
  CrawlerName = "string"
)
}
}

