% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_connection}
\alias{update_connection}
\title{Updates a connection definition in the Data Catalog}
\usage{
update_connection(Name, ConnectionInput, CatalogId = NULL)
}
\arguments{
\item{Name}{The name of the connection definition to update.}

\item{ConnectionInput}{A \code{ConnectionInput} object that redefines the connection in question.}

\item{CatalogId}{The ID of the Data Catalog in which the connection resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Updates a connection definition in the Data Catalog.
}
\examples{

}
