% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_crawler_schedule}
\alias{update_crawler_schedule}
\title{Updates the schedule of a crawler using a cron expression}
\usage{
update_crawler_schedule(CrawlerName, Schedule = NULL)
}
\arguments{
\item{CrawlerName}{[required] Name of the crawler whose schedule to update.}

\item{Schedule}{The updated \code{cron} expression used to specify the schedule (see \href{http://docs.aws.amazon.com/glue/latest/dg/monitor-data-warehouse-schedule.html}{Time-Based Schedules for Jobs and Crawlers}. For example, to run something every day at 12:15 UTC, you would specify: \code{cron(15 12 * * ? *)}.}
}
\description{
Updates the schedule of a crawler using a \code{cron} expression.
}
\section{Accepted Parameters}{
\preformatted{update_crawler_schedule(
  CrawlerName = "string",
  Schedule = "string"
)
}
}

