% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_job}
\alias{update_job}
\title{Updates an existing job definition}
\usage{
update_job(JobName, JobUpdate)
}
\arguments{
\item{JobName}{Name of the job definition to update.}

\item{JobUpdate}{Specifies the values with which to update the job definition.}
}
\description{
Updates an existing job definition.
}
\examples{

}
