% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_partition}
\alias{update_partition}
\title{Updates a partition}
\usage{
update_partition(DatabaseName, TableName, PartitionValueList, PartitionInput,
  CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database in which the table in question resides.}

\item{TableName}{The name of the table where the partition to be updated is located.}

\item{PartitionValueList}{A list of the values defining the partition.}

\item{PartitionInput}{The new partition object to which to update the partition.}

\item{CatalogId}{The ID of the Data Catalog where the partition to be updated resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Updates a partition.
}
\examples{

}
