% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_trigger}
\alias{update_trigger}
\title{Updates a trigger definition}
\usage{
update_trigger(Name, TriggerUpdate)
}
\arguments{
\item{Name}{The name of the trigger to update.}

\item{TriggerUpdate}{The new values with which to update the trigger.}
}
\description{
Updates a trigger definition.
}
\examples{

}
