% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_user_defined_function}
\alias{update_user_defined_function}
\title{Updates an existing function definition in the Data Catalog}
\usage{
update_user_defined_function(DatabaseName, FunctionName, FunctionInput,
  CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database where the function to be updated is located.}

\item{FunctionName}{The name of the function.}

\item{FunctionInput}{A \code{FunctionInput} object that re-defines the function in the Data Catalog.}

\item{CatalogId}{The ID of the Data Catalog where the function to be updated is located. If none is supplied, the AWS account ID is used by default.}
}
\description{
Updates an existing function definition in the Data Catalog.
}
\examples{

}
