% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{batch_delete_connection}
\alias{batch_delete_connection}
\title{Deletes a list of connection definitions from the Data Catalog}
\usage{
batch_delete_connection(ConnectionNameList, CatalogId = NULL)
}
\arguments{
\item{ConnectionNameList}{A list of names of the connections to delete.}

\item{CatalogId}{The ID of the Data Catalog in which the connections reside. If none is supplied, the AWS account ID is used by default.}
}
\description{
Deletes a list of connection definitions from the Data Catalog.
}
\examples{

}
