% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_connection}
\alias{create_connection}
\title{Creates a connection definition in the Data Catalog}
\usage{
create_connection(ConnectionInput, CatalogId = NULL)
}
\arguments{
\item{ConnectionInput}{A \code{ConnectionInput} object defining the connection to create.}

\item{CatalogId}{The ID of the Data Catalog in which to create the connection. If none is supplied, the AWS account ID is used by default.}
}
\description{
Creates a connection definition in the Data Catalog.
}
\examples{

}
