% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{create_security_configuration}
\alias{create_security_configuration}
\title{Creates a new security configuration}
\usage{
create_security_configuration(Name, EncryptionConfiguration)
}
\arguments{
\item{Name}{The name for the new security configuration.}

\item{EncryptionConfiguration}{The encryption configuration for the new security configuration.}
}
\description{
Creates a new security configuration.
}
\examples{

}
