% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_classifier}
\alias{delete_classifier}
\title{Removes a classifier from the Data Catalog}
\usage{
delete_classifier(Name)
}
\arguments{
\item{Name}{Name of the classifier to remove.}
}
\description{
Removes a classifier from the Data Catalog.
}
\examples{

}
