% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_connection}
\alias{delete_connection}
\title{Deletes a connection from the Data Catalog}
\usage{
delete_connection(
  CatalogId = "string",
  ConnectionName = "string"
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connection resides. If none is supplied, the AWS account ID is used by default.}

\item{ConnectionName}{[required] The name of the connection to delete.}
}
\description{
Deletes a connection from the Data Catalog.
}
\examples{

}
