% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_dev_endpoint}
\alias{delete_dev_endpoint}
\title{Deletes a specified DevEndpoint}
\usage{
delete_dev_endpoint(
  EndpointName = "string"
)
}
\arguments{
\item{EndpointName}{[required] The name of the DevEndpoint.}
}
\description{
Deletes a specified DevEndpoint.
}
\examples{

}
