% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_partition}
\alias{delete_partition}
\title{Deletes a specified partition}
\usage{
delete_partition(DatabaseName, TableName, PartitionValues, CatalogId = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database in which the table in question resides.}

\item{TableName}{The name of the table where the partition to be deleted is located.}

\item{PartitionValues}{The values that define the partition.}

\item{CatalogId}{The ID of the Data Catalog where the partition to be deleted resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Deletes a specified partition.
}
\examples{

}
