% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{delete_table_version}
\alias{delete_table_version}
\title{Deletes a specified version of a table}
\usage{
delete_table_version(CatalogId = NULL, DatabaseName, TableName,
  VersionId)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The database in the catalog in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{TableName}{[required] The name of the table. For Hive compatibility, this name is entirely lowercase.}

\item{VersionId}{[required] The ID of the table version to be deleted. A \code{VersionID} is a string representation of an integer. Each version is incremented by 1.}
}
\description{
Deletes a specified version of a table.
}
\section{Accepted Parameters}{
\preformatted{delete_table_version(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  VersionId = "string"
)
}
}

