% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_connection}
\alias{get_connection}
\title{Retrieves a connection definition from the Data Catalog}
\usage{
get_connection(Name, CatalogId = NULL)
}
\arguments{
\item{Name}{The name of the connection definition to retrieve.}

\item{CatalogId}{The ID of the Data Catalog in which the connection resides. If none is supplied, the AWS account ID is used by default.}
}
\description{
Retrieves a connection definition from the Data Catalog.
}
\examples{

}
