% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_dev_endpoint}
\alias{get_dev_endpoint}
\title{Retrieves information about a specified DevEndpoint}
\usage{
get_dev_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] Name of the DevEndpoint for which to retrieve information.}
}
\description{
Retrieves information about a specified DevEndpoint.
}
\details{
When you create a development endpoint in a virtual private cloud (VPC), AWS Glue returns only a private IP address, and the public IP address field is not populated. When you create a non-VPC development endpoint, AWS Glue returns only a public IP address.
}
\section{Accepted Parameters}{
\preformatted{get_dev_endpoint(
  EndpointName = "string"
)
}
}

