% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_mapping}
\alias{get_mapping}
\title{Creates mappings}
\usage{
get_mapping(
  Source = list(
    DatabaseName = "string",
    TableName = "string"
  ),
  Sinks = list(
    list(
      DatabaseName = "string",
      TableName = "string"
    )
  ),
  Location = list(
    Jdbc = list(
      list(
        Name = "string",
        Value = "string",
        Param = TRUE|FALSE
      )
    ),
    S3 = list(
      list(
        Name = "string",
        Value = "string",
        Param = TRUE|FALSE
      )
    ),
    DynamoDB = list(
      list(
        Name = "string",
        Value = "string",
        Param = TRUE|FALSE
      )
    )
  )
)
}
\arguments{
\item{Source}{[required] Specifies the source table.}

\item{Sinks}{A list of target tables.}

\item{Location}{Parameters for the mapping.}
}
\description{
Creates mappings.
}
\examples{

}
