% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_partition}
\alias{get_partition}
\title{Retrieves information about a specified partition}
\usage{
get_partition(CatalogId = NULL, DatabaseName, TableName, PartitionValues)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partition in question resides. If none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partition resides.}

\item{TableName}{[required] The name of the partition's table.}

\item{PartitionValues}{[required] The values that define the partition.}
}
\description{
Retrieves information about a specified partition.
}
\section{Accepted Parameters}{
\preformatted{get_partition(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  PartitionValues = list(
    "string"
  )
)
}
}

