% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{get_security_configurations}
\alias{get_security_configurations}
\title{Retrieves a list of all security configurations}
\usage{
get_security_configurations(
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Retrieves a list of all security configurations.
}
\examples{

}
