% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{stop_trigger}
\alias{stop_trigger}
\title{Stops a specified trigger}
\usage{
stop_trigger(Name)
}
\arguments{
\item{Name}{The name of the trigger to stop.}
}
\description{
Stops a specified trigger.
}
\examples{

}
