% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_table}
\alias{update_table}
\title{Updates a metadata table in the Data Catalog}
\usage{
update_table(DatabaseName, TableInput, CatalogId = NULL, SkipArchive = NULL)
}
\arguments{
\item{DatabaseName}{The name of the catalog database in which the table resides. For Hive compatibility, this name is entirely lowercase.}

\item{TableInput}{An updated \code{TableInput} object to define the metadata table in the catalog.}

\item{CatalogId}{The ID of the Data Catalog where the table resides. If none is supplied, the AWS account ID is used by default.}

\item{SkipArchive}{By default, \code{UpdateTable} always creates an archived version of the table before updating it. If \code{skipArchive} is set to true, however, \code{UpdateTable} does not create the archived version.}
}
\description{
Updates a metadata table in the Data Catalog.
}
\examples{

}
