% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glue_operations.R
\name{update_user_defined_function}
\alias{update_user_defined_function}
\title{Updates an existing function definition in the Data Catalog}
\usage{
update_user_defined_function(CatalogId = NULL, DatabaseName,
  FunctionName, FunctionInput)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the function to be updated is located. If none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the function to be updated is located.}

\item{FunctionName}{[required] The name of the function.}

\item{FunctionInput}{[required] A \code{FunctionInput} object that re-defines the function in the Data Catalog.}
}
\description{
Updates an existing function definition in the Data Catalog.
}
\section{Accepted Parameters}{
\preformatted{update_user_defined_function(
  CatalogId = "string",
  DatabaseName = "string",
  FunctionName = "string",
  FunctionInput = list(
    FunctionName = "string",
    ClassName = "string",
    OwnerName = "string",
    OwnerType = "USER"|"ROLE"|"GROUP",
    ResourceUris = list(
      list(
        ResourceType = "JAR"|"FILE"|"ARCHIVE",
        Uri = "string"
      )
    )
  )
)
}
}

